refactor: bullet/strip_ansi 로직을 osc.rs에서 claude_bullet.rs로 분리
osc.rs는 OSC 시퀀스 파싱 전담, claude_bullet.rs는 Claude Code 출력 해석 전담으로
단일 책임 원칙에 맞게 모듈을 분리한다. 추가 개선:
- 매직넘버 500 → BULLET_MESSAGE_SCAN_BYTES 상수화
- unwrap() → if let Some(ch) 안전 처리
- OSC ST terminator (ESC \) 핸들링 개선
- Claude Code 종료 시 claude_message 클리어 로직 추가
