+++ ui/src/hooks/useSyncEvents.test.ts
@@ -289,6 +289,28 @@ describe("useSyncEvents", () => {
     expect(instance?.activity).toEqual({ type: "interactiveApp", name: "Claude" });
   });
 
+  it("registers claude-message-changed listener on mount", () => {
+    renderHook(() => useSyncEvents());
+    expect(mockOnClaudeMessageChanged).toHaveBeenCalledWith(expect.any(Function));
+  });
+
+  it("updates terminal claudeMessage on claude-message-changed event", () => {
+    useTerminalStore.getState().registerInstance({
+      id: "t1",
+      profile: "WSL",
+      syncGroup: "g1",
+      workspaceId: "ws-1",
+    });
+
+    renderHook(() => useSyncEvents());
+
+    const callback = mockOnClaudeMessageChanged.mock.calls[0][0];
+    callback({ terminalId: "t1", message: "모든 테스트 통과했습니다." });
+
+    const instance = useTerminalStore.getState().instances.find((i) => i.id === "t1");
+    expect(instance?.claudeMessage).toBe("모든 테스트 통과했습니다.");
+  });
+
   it("calls markClaudeTerminal when command text detects Claude", () => {
     useTerminalStore.getState().registerInstance({
       id: "t1",
